LineChartコンポーネントの削除
LineChartは1箇所しか使用していないため
