--- src/components/commons/ui/LineChart/index.tsx
@@ -1,1 +1,0 @@
-export * from './LineChart'
