+++ src/components/features/mulliganCalculator/MulliganLineChart/MulliganLineChart.tsx
@@ -1,6 +1,30 @@
+'use client'
+
 import React from 'react'
+import { Line } from 'react-chartjs-2'
+import {
+  Chart as ChartJS,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  Title,
+  Tooltip,
+  Legend,
+  ChartOptions,
+  ChartData,
+} from 'chart.js'
 import { calculateMulliganProbability } from './MulliganLineChart.hooks'
-import { LineChart } from '@components/commons/ui/LineChart'
+
+ChartJS.register(
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  Title,
+  Tooltip,
+  Legend
+)
 
 const labels = [
   'マリガン後',
@@ -13,6 +37,7 @@ interface MulliganLineChartProps {
   renCount: number
   mulliganCount: number
   wantCardCount: number
+  height?: number
 }
 
 export const MulliganLineChart: React.FC<MulliganLineChartProps> = ({
@@ -21,6 +46,7 @@ export const MulliganLineChart: React.FC<MulliganLineChartProps> = ({
   renCount,
   mulliganCount,
   wantCardCount,
+  height = 300,
 }) => {
   const probabilities = calculateMulliganProbability({
     deckSize,
@@ -30,13 +56,46 @@ export const MulliganLineChart: React.FC<MulliganLineChartProps> = ({
     mulliganCount,
   })
 
-  return (
-    <LineChart
-      labels={labels}
-      lineData={probabilities}
-      xText="ドロー枚数"
-      yMin={Math.max(Math.floor(probabilities[0] - 5), 0)}
-      yText="1枚以上引く確率 (%)"
-    />
-  )
+  const data: ChartData<'line'> = {
+    labels,
+    datasets: [
+      {
+        label: '1枚以上引く確率 (%)',
+        data: probabilities,
+        borderColor: '#5EB1EF',
+        backgroundColor: 'rgba(94, 177, 239, 0.2)',
+        borderWidth: 2,
+        pointRadius: 3,
+      },
+    ],
+  }
+
+  const options: ChartOptions<'line'> = {
+    responsive: true,
+    maintainAspectRatio: false,
+    scales: {
+      y: {
+        beginAtZero: false,
+        min: Math.max(Math.floor(probabilities[0] - 5), 0),
+        suggestedMax: 100,
+        ticks: {
+          callback: function (value) {
+            return value + '%'
+          },
+        },
+      },
+      x: {
+        title: {
+          display: true,
+          text: 'ドロー枚数',
+          padding: {
+            top: -24,
+            bottom: 0,
+          },
+        },
+      },
+    },
+  }
+
+  return <Line data={data} height={height} options={options} />
 }
